Flipped TerminusCollection; defaults to fetchless; added APICollection
