+++ src/Collections/SSHKeys.php
@@ -64,13 +64,11 @@ class SSHKeys extends UserOwnedCollection
     /**
      * Fetches model data from API and instantiates its model instances
      *
-     * @param array $options params to pass to url request
      * @return SSHKeys $this
      */
-    public function fetch(array $options = [])
+    public function fetch()
     {
-        $data = isset($options['data']) ? $options['data'] : $this->getCollectionData($options);
-        if (!is_null($data)) {
+        if (!is_null($data = $this->getData())) {
             foreach ($data as $uuid => $ssh_key) {
                 $model_data = (object)['id' => $uuid, 'key' => $ssh_key,];
                 $this->add($model_data);
