+++ src/Collections/Tags.php
@@ -9,7 +9,7 @@ use Pantheon\Terminus\Models\Tag;
  * Class Tags
  * @package Pantheon\Terminus\Collections
  */
-class Tags extends TerminusCollection
+class Tags extends APICollection
 {
     public static $pretty_name = 'tags';
     /**
