+++ src/Collections/WorkflowOperations.php
@@ -8,7 +8,7 @@ use Pantheon\Terminus\Models\WorkflowOperation;
  * Class WorkflowOperations
  * @package Pantheon\Terminus\Collections
  */
-class WorkflowOperations extends FetchlessCollection
+class WorkflowOperations extends TerminusCollection
 {
     public static $pretty_name = 'workflow operations';
     /**
